requirements.txt
annotated-types==0.7.0
anyio==4.6.0
beautifulsoup4==4.12.3
certifi==2024.8.30
charset-normalizer==3.4.0
click==8.1.7
colorlog==6.8.2
dnspython==2.7.0
email_validator==2.2.0
fastapi==0.110.0
frozendict==2.4.6
gevent==24.10.3
greenlet==3.1.1
h11==0.14.0
html5lib==1.1
httpcore==1.0.6
httpx==0.27.2
idna==3.10
iniconfig==2.0.0
kafka-python @ git+https://github.com/dpkp/kafka-python.git@a6d0579d3cadd3826dd364b01bc12a2173139abc
lxml==5.3.0
multitasking==0.0.11
mysql-connector-python==9.1.0
numpy==2.1.2
packaging==24.1
pandas==2.2.3
passlib==1.7.4
peewee==3.17.7
platformdirs==4.3.6
pluggy==1.5.0
pydantic==2.9.2
pydantic_core==2.23.4
pytest==8.3.3
pytest-asyncio==0.24.0
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
pytz==2024.2
redis==5.1.1
requests==2.32.3
setuptools==75.2.0
six==1.16.0
sniffio==1.3.1
soupsieve==2.6
starlette==0.36.3
typing_extensions==4.12.2
tzdata==2024.2
urllib3==2.2.3
uvicorn==0.31.0
webencodings==0.5.1
websocket==0.2.1
websocket-client==1.8.0
yfinance==0.2.45
zope.event==5.0
zope.interface==7.1.0
